refactor(traces): move default log-store construction into the app layer
The C2 wiring landed the ClickHouse-repo default construction (isClickHouseEnabled,
the CH resolver, the repo classes) inside the legacy src/server/traces/trace.service.ts.
Move it to an app-layer factory createDefaultLogRecordStorageService() so the legacy
layer holds as little new code as possible and never touches ClickHouse wiring; the
lazy default there just delegates to the factory. Also: build the default lazily (so
constructing TraceService in unit tests never needs to mock CH), and null-safe the
coding-agent origin gate (trace.metadata may be undefined).
